Status Submenu: separate [Human*] from helper into [human.rs]
